requirements.txt
Django==2.1
django-tinymce==2.7.0
pytz==2018.5
